refactor: 💡 rename the theme context to global context
✅ Closes: #157
