+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "src",
-  "version": "0.1.0+45.4",
+  "version": "0.1.0+45.5",
   "private": true,
   "scripts": {
     "dev": "next dev",
