+++ src/components/market_section/market_section.tsx
@@ -1,7 +1,7 @@
 import {useState, useContext} from 'react';
 import TradingHeader from '../trading_header/trading_header';
 import TradingView from '../trading_view/trading_view';
-import {GlobalContext, useGlobal} from '../../lib/contexts/theme_context';
+import {GlobalContext, useGlobal} from '../../lib/contexts/global_context';
 import TradingViewMobile from '../trading_view_mobile/trading_view_mobile';
 import TradingHeaderMobile from '../trading_header_mobile/trading_header_mobile';
 import TradeStatistics from '../trade_statistics/trade_statistics';
