+++ src/components/trade_page_body/trade_page_body.tsx
@@ -4,7 +4,7 @@ import Footer from '../footer/footer';
 import MarketSection from '../market_section/market_section';
 import {MarketProvider} from '../../lib/contexts/market_context';
 import {UserProvider} from '../../lib/contexts/user_context';
-import {GlobalContext, useGlobal} from '../../lib/contexts/theme_context';
+import {GlobalContext, useGlobal} from '../../lib/contexts/global_context';
 import OrderSectionMobile from '../order_section_mobile/order_section_mobile';
 
 const TradePageBody = () => {
