+++ src/lib/contexts/global_context.tsx
@@ -1,21 +1,54 @@
-import {createContext, useState, useEffect, useContext} from 'react';
-
+import {createContext, useState, useEffect, useContext, Dispatch, SetStateAction} from 'react';
+import useWindowSize from '../hooks/use_window_size';
+import {LAYOUT_BREAKPOINT} from '../../constants/display';
 export interface IGlobalProvider {
   children: React.ReactNode;
 }
 
+export type LayoutAssertionUnion = 'mobile' | 'desktop';
+export type ColorModeUnion = 'light' | 'dark';
+
 export interface IGlobalContext {
-  toast: string;
+  width: number;
+  height: number;
+  layoutAssertion: LayoutAssertionUnion;
+  initialColorMode: ColorModeUnion;
+  colorMode: ColorModeUnion;
+  toggleColorMode: () => void;
 }
 
 export const GlobalContext = createContext<IGlobalContext>({
-  toast: '',
+  width: 0,
+  height: 0,
+  layoutAssertion: '' as LayoutAssertionUnion,
+  initialColorMode: '' as ColorModeUnion,
+  colorMode: '' as ColorModeUnion,
+  toggleColorMode: () => null,
+  // setColorMode: (() => {}) as Dispatch<SetStateAction<ColorModeUnion>>,
 });
 
-const toast = 'dummy toast';
-
 export const GlobalProvider = ({children}: IGlobalProvider) => {
-  const defaultValue = {toast};
+  const windowSize = useWindowSize();
+  const {width, height} = windowSize;
+  const layoutAssertion: LayoutAssertionUnion = width < LAYOUT_BREAKPOINT ? 'mobile' : 'desktop';
 
+  const initialColorMode: ColorModeUnion = 'dark';
+  const [colorMode, setColorMode] = useState<ColorModeUnion>(initialColorMode);
+  const toggleColorMode = () => {
+    setColorMode(colorMode === 'light' ? 'dark' : 'light');
+  };
+
+  const defaultValue = {
+    width,
+    height,
+    layoutAssertion,
+    initialColorMode,
+    colorMode,
+    toggleColorMode,
+  };
   return <GlobalContext.Provider value={defaultValue}>{children}</GlobalContext.Provider>;
+};
+
+export const useGlobal = () => {
+  return useContext(GlobalContext);
 };
