+++ src/pages/_app.tsx
@@ -5,7 +5,7 @@ import type {AppProps} from 'next/app';
 import {appWithTranslation} from 'next-i18next';
 import {MarketProvider} from '../lib/contexts/market_context';
 import {UserProvider} from '../lib/contexts/user_context';
-import {GlobalProvider} from '../lib/contexts/theme_context';
+import {GlobalProvider} from '../lib/contexts/global_context';
 import {ToastContainer} from 'react-toastify';
 
 function App({Component, pageProps}: AppProps) {
